p94写商品管理中。。。。。

+++ gulimall-product/src/main/java/com/athl/gulimall/product/entity/vo/Attr.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @Data
 public class Attr {
 
-    private int attrId;
+    private Long attrId;
     private String attrName;
     private String attrValue;
