+++ gulimall-product/src/main/java/com/athl/gulimall/product/entity/vo/BaseAttrs.java
@@ -14,8 +14,11 @@ import lombok.Data;
 @Data
 public class BaseAttrs {
 
-    private int attrId;
+    private Long attrId;
     private String attrValues;
+    /**
+     * 是否快速展示
+     */
     private int showDesc;
 
 }
